requirements.txt
psutil
boto3
numpy
Pillow
pyautogui
pynput